+++ src/lib/core/LightboxClient/index.ts
@@ -1,3 +1,4 @@
+import { redirectBasedOnDriver } from '@/helpers';
 import { InitOptions, LightboxInstance } from '@/types';
 import { mountIFrameElement, mountListener } from './assemblers';
 
@@ -5,8 +6,10 @@ export const LightboxClient = {
     init: (url: string, options?: InitOptions): LightboxInstance => {
         const lightbox: LightboxInstance = {
             callbacks: options?.callbacks ?? {},
+            allowRedirects: options?.allowRedirects ?? true,
             styles: options?.styles ?? {},
             open: () => {
+                if (lightbox.allowRedirects) redirectBasedOnDriver(url);
                 mountListener(lightbox.callbacks);
                 mountIFrameElement(url, lightbox.styles);
             },