+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- AllowRedirects option
+- Helpers module
 - Height and width option
 - Dimensions option
 - Defined configuration file